[MINOR][SS] Reuse KafkaSourceInitialOffsetWriter to deduplicate
