+++ external/kafka-0-10-sql/src/main/scala/org/apache/spark/sql/kafka010/KafkaSource.scala
@@ -98,37 +98,7 @@ private[kafka010] class KafkaSource(
    * `KafkaConsumer.poll` may hang forever (KAFKA-1894).
    */
   private lazy val initialPartitionOffsets = {
-    val metadataLog =
-      new HDFSMetadataLog[KafkaSourceOffset](sqlContext.sparkSession, metadataPath) {
-        override def serialize(metadata: KafkaSourceOffset, out: OutputStream): Unit = {
-          out.write(0) // A zero byte is written to support Spark 2.1.0 (SPARK-19517)
-          val writer = new BufferedWriter(new OutputStreamWriter(out, StandardCharsets.UTF_8))
-          writer.write("v" + VERSION + "\n")
-          writer.write(metadata.json)
-          writer.flush
-        }
-
-        override def deserialize(in: InputStream): KafkaSourceOffset = {
-          in.read() // A zero byte is read to support Spark 2.1.0 (SPARK-19517)
-          val content = IOUtils.toString(new InputStreamReader(in, StandardCharsets.UTF_8))
-          // HDFSMetadataLog guarantees that it never creates a partial file.
-          assert(content.length != 0)
-          if (content(0) == 'v') {
-            val indexOfNewLine = content.indexOf("\n")
-            if (indexOfNewLine > 0) {
-              validateVersion(content.substring(0, indexOfNewLine), VERSION)
-              KafkaSourceOffset(SerializedOffset(content.substring(indexOfNewLine + 1)))
-            } else {
-              throw new IllegalStateException(
-                s"Log file was malformed: failed to detect the log file version line.")
-            }
-          } else {
-            // The log was generated by Spark 2.1.0
-            KafkaSourceOffset(SerializedOffset(content))
-          }
-        }
-      }
-
+    val metadataLog = new KafkaSourceInitialOffsetWriter(sqlContext.sparkSession, metadataPath)
     metadataLog.get(0).getOrElse {
       val offsets = startingOffsets match {
         case EarliestOffsetRangeLimit => KafkaSourceOffset(kafkaReader.fetchEarliestOffsets())
